refactor: move config to constant
